quantitative_eval/requirements_consolidated.txt
# Romeo & Juliet RAG系统 - 整合的Python依赖文件
# 基于项目实际使用情况和Python 3.12兼容性

# ============================================================================
# 核心机器学习和NLP包
# ============================================================================
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.1.0
transformers>=4.30.0
tokenizers>=0.14.0

# ============================================================================
# 现代向量检索（Python 3.12兼容）
# ============================================================================
faiss-cpu>=1.12.0                  # 现代FAISS语义检索
sentence-transformers>=5.1.0        # 现代文本编码

# ============================================================================
# 传统检索和评估
# ============================================================================
pyserini>=0.22.0                   # 传统BM25检索
ranx>=0.3.18                       # 检索评估工具
pytrec-eval>=0.5                   # TREC评估指标

# ============================================================================
# 文本处理和NLP
# ============================================================================
spacy>=3.7.0                       # 文本处理
beautifulsoup4>=4.12.0             # HTML解析
lxml>=4.9.0                        # XML处理

# ============================================================================
# 语音和多媒体（用于完整语音RAG系统）
# ============================================================================
openai-whisper>=20231117            # 语音转文字（现代语音RAG必需）
gTTS>=2.3.0                         # 文字转语音（现代语音RAG必需）
sounddevice>=0.4.0                  # 音频录制（现代语音RAG必需）
pydub>=0.25.0                       # 音频处理（现代语音RAG必需）
pygame>=2.0.0                       # 音频播放（现代语音RAG必需）
scipy>=1.10.0                       # 音频信号处理（sounddevice依赖）

# ============================================================================
# 数据处理和工具
# ============================================================================
tqdm>=4.66.0                       # 进度条
requests>=2.31.0                   # HTTP请求
pyyaml>=6.0.0                      # YAML配置文件
click>=8.1.0                       # 命令行工具
joblib>=1.3.0                      # 并行处理
packaging>=23.0                    # 包管理

# ============================================================================
# 可视化和分析
# ============================================================================
matplotlib>=3.7.0                  # 绘图
seaborn>=0.13.0                    # 统计绘图

# ============================================================================
# Hugging Face生态
# ============================================================================
huggingface-hub>=0.16.0            # 模型下载
safetensors>=0.3.0                 # 安全张量格式

# ============================================================================
# 评估和NLG指标（如果需要详细评估）
# ============================================================================
nltk>=3.8.0                        # 自然语言工具包（文本生成评估）
rouge>=1.0.0                       # ROUGE评估（文本生成评估）
bert-score>=0.3.0                  # BERTScore评估（文本生成评估）

# ============================================================================
# 说明
# ============================================================================
#
# 核心系统（必需）：
# - 使用上述未注释的包即可运行现代化FAISS+RAG系统
# - 总计约25个核心依赖包
#
# 扩展功能（可选）：
# - 注释掉的包用于完整的语音RAG系统和详细评估
# - 如需要语音交互功能，请取消注释语音相关包
# - 如需要详细的NLG评估，请取消注释评估相关包
#
# 安装方式：
# pip install -r requirements_consolidated.txt
#
# 与旧版本对比：
# - 旧requirements.txt包含109个包，许多是不必要的传递依赖
# - 新版本只包含25个核心包，专注于项目实际需求
# - 移除了不兼容Python 3.12的包（如nmslib）
# - 升级了关键包（FAISS 1.7.4 → 1.12.0）